Using short instead of byte for Load Priority
I don't like bytes

+++ src/com/thebubblenetwork/api/framework/plugin/LoadPriority.java
@@ -4,5 +4,5 @@ package com.thebubblenetwork.api.framework.plugin;
  * Created by Jacob on 12/12/2015.
  */
 public class LoadPriority {
-    public static byte LOWEST = 127, HIGHEST = 1, NORMAL = 64, LOW = 96, HIGH = 32;
+    public static short LOWEST = 127, HIGHEST = 1, NORMAL = 64, LOW = 96, HIGH = 32;
 }
